requirements.txt
Flask==0.10.1

# Hello. This is Nick from the future (March 2019 to be exact). An emergency
# addition had to be made below to fix an issue related to a recent version of
# werkzeug. This library was not version locked in this file.
#
# The line below isn't covered on video, but it locks werkzeug to the latest
# version that works with this course's code base. A future update video will
# cover upgrading this package and more.
werkzeug==0.14.1

# Application server for both development and production.
gunicorn==19.4.5

# Testing and static analysis.
#
# Hello. This is Nick from the future (December 2019 to be exact). Since we
# upgraded to Python 3.7.x, we also have to update pytest to 5.x.x since older
# versions of it are not compatable with Python 3.7+.
#
# This line is different than what's on video but we do cover this update in
# more detail in the October 2019 update video. When you run your tests you
# may also see extra warnings and details than what's on video. Don't sweat it.
pytest==5.1.0
pytest-cov==2.7.1
flake8==3.7.8

# CLI.
Click==6.4

# Data and workers.
#
# Hello. This is Nick from the future (August 2019 to be exact). An emergency
# addition had to be made below to fix an issue related to Docker recently
# changing the internals of their Python image.
#
# On video it shows psycopg2 version 2.6.1 but I had to bump it to 2.8.3 here.
psycopg2==2.8.3
Flask-SQLAlchemy==2.1

# Hello. This is Nick from the future (March 2019 to be exact). An emergency
# addition had to be made below to fix an issue that resulted in SQLAlchemy
# throwing an error. Flask-SQLAlchemy 2.1 doesn't version lock SQLAlchemy and
# SQLAlchemy 1.3.x introduced breaking changes.
#
# The line below isn't covered on video, but it locks SQLAlchemy to 1.2.0. A
# future update video will cover upgrading SQLAlchemy to 1.3.0 and more.
SQLAlchemy==1.2.0

# Hello. This is Nick from the future (December 2019 to be exact). Since we
# upgraded to Python 3.7.x, we also have to update Celery to 4.3.x since older
# versions of it are not compatible with Python 3.7+.
#
# This line is different than what's on video but we do cover this update in
# more detail in the October 2019 update video.
#
# And this is Nick again from September 2020. Turns out Celery 4.3.x started to
# break due to one of its dependencies not being locked. Now we're using 4.4.0.
redis==3.3.7
celery==4.4.0

# Forms.
Flask-WTF==0.9.5
WTForms-Components==0.9.7

# Utils.
fake-factory==0.5.7

# Extensions.
flask-debugtoolbar==0.10.0
Flask-Mail==0.9.1
Flask-Login==0.3.2

ipfshttpclient==0.7.0
pybase62